+++ forui/CHANGELOG.md
@@ -28,6 +28,7 @@
 * Add `animationTween` to `FTappable.animated(...)`.
 * Add `FValueNotifier.addValueListener(...)`.
 * Add `FValueNotifier.removeValueListener(...)`.
+* Add `buildCounter` to `FTextField`.
 
 ### Changes